+++ cosmosid/commands/artifacts.py
@@ -28,7 +28,7 @@ class Artifacts(Lister):
         parser.add_argument(
             "--type",
             "-t",
-            choices=["fastqc-zip"],
+            choices=["fastqc-zip", "champ-supplementary"],
             type=str,
             help="Artifact type to download",
         )