+++ cosmosid/commands/runs.py
@@ -37,14 +37,22 @@ class Runs(Lister):
         """get json with runs for a file and prepare for output."""
         ids = utils.key_len(parsed_args.id, "ID")
         runs = self.app.cosmosid.sample_run_list(ids)
-        header = ["id", "status", "created"]
+        header = [
+            "id",
+            "created",
+            "workflow_name",
+            "workflow_version",
+            "status",
+            "artifact_types",
+        ]
+        for_output = [["" for _ in header]]
         if runs:
             if not runs["runs"]:
                 self.app.logger.info(
                     "\nThere are no runs for sample %s (id: %s)", runs["file_name"], ids
                 )
 
-                for_output = [[" ", " ", " "]]
+                for_output = [["" for i in header]]
                 return (header, for_output)
         else:
             raise Exception("Exception occurred.")
@@ -73,6 +81,10 @@ class Runs(Lister):
             "id": ["id", "str", _del_none],
             "status": ["status", "str", _del_none],
             "created": ["created", "str", _set_date],
+            "workflow_id": ["workflow_id", "str", _del_none],
+            "workflow_name": ["workflow_name", "str", _del_none],
+            "workflow_version": ["workflow_version", "str", _del_none],
+            "artifact_types": ["artifact_types", "str", _del_none],
         }
 
         # we need just runs for output
@@ -86,9 +98,21 @@ class Runs(Lister):
                     key=itemgetter(field_maps[parsed_args.order.lower()][0]),
                     reverse=(not parsed_args.up),
                 )
-        for_output = [
-            [item[field_maps[field][0]] for field in header] for item in runs_data
-        ]
+        for_output = []
+        for item in runs_data:
+
+            items = []
+            for field in header:
+                name = field_maps[field][0]
+                i = item[name]
+                if name == "artifact_types":
+                    artifacts = i.split(",")
+                    i = "".join([f"{a} \n" for a in artifacts])
+
+                items.append(i)
+
+            for_output.append(items)
+
         self.app.logger.info(
             "\nRuns list for file %s (id: %s)", runs["file_name"], ids)
         return (header, for_output)